Fixed Sigmoid and TanH

+++ tools/SeeDot/seedot/Predictor/datatypes.h
@@ -5,7 +5,7 @@
 
 #define INT16
 typedef int32_t ACINT;
-typedef int16_t MYINT;
+typedef int8_t MYINT;
 typedef int16_t MYITE;
 typedef uint16_t MYUINT;
